s/blob/artifact, handle all blobs at the job level
This is a breaking change to the backup metadata format. Previously, custom
named artifacts were called blobs, and were the only artifacts handled at the
job level of granularity.

By handling all artifacts at the same level, we open the door to simplifying
much of the code, as well as making the metadata format more consistent and
logical.

There is almost certainly a lot of code that can now be simplified or refactored
as a result of this change. We can revisit this, but for now we wanted to get to
a point where the metadata format was stable before shipping. The only thing we
may still change in this area that would break compatibility is the file names
and directory structure for artifacts on disk.

Signed-off-by: Will Pragnell <wpragnell@pivotal.io>
Signed-off-by: Derik Evangelista <devangelista@pivotal.io>