Limit parallelism to avoid SSH issues
Trying to perform too many SSH connections at the same time does not
work because of 2 reasons:

* By default, SSH will only authenticate 10 connections at a time, and
  drop any other connection (see the MaxStartups config parameter);
* Even after increasing MaxStartups, SSH will still struggle to handle
  too many simultaneous connection attempts, so it will drop some.

This commit changes the ParallelExecutor to only run a maximum number of
Executables at any time. The number is stored in the maxInFlight field
and currently hardcoded to 10 in the constructor, to match the default
value of MaxStartups.

[#156816267]

Signed-off-by: Emmanouil Kiagias <ekiagias@pivotal.io>