Bump bosh-cli in dependencies
We had to bump bosh-cli to consume a bosh_all_proxy fix. For the bump to
work, we had to bump, bosh-agent and bosh-utils as well. Since we hadn't
bump for quite a while, there were some changes in function signatures
so this commit includes all the necessary changes for the bump to work.

[#162789201]