Fix and clarify flakey test
As far as I can tell, this test is intended to check that long-running
processes don't get accidentally killed by SSH timeouts.

I've updated the BDD language to try to make that clearer.

Also, the previous version of this test appeared to try to use a count
of the goroutines currently running in the go runtime to detect when a
remote process had finished running. It might have been doing
something more clever than that, but that's my best guess.

Assuming I'm right, I don't see any reason to do it that way. Instead
we can just wait for the expected output when the process finishes.

The goroutine thing was fragile. When running the test in isolation it
always passed, but when running as part of a ginkgo suite, it often
failed. I suspect that the number of goroutines ginkgo itself uses may
have changed since the test was first written.