[ci] merge promote and github release jobs
Previously the promote job was creating a draft release on Github and
then pushing the promoted artefacts there. Later on the
publish-to-github job would just add the release notes and make the
release public. It turns out the github-release resource would not keep
the artefacts already uploaded in the draft release, and would remove
them instead. This made it impossible to use Github draft releases as a
storage for promoted artefacts.

[#179484418]