Make integration tests work with recent docker versions
The function mockInstance.Address() returns the address that you can
use to SSH into a particular docker container. To do this, it needs to
discover the port on your local machine that docker has mapped to port
22 inside the container.

The "docker port" command provides this information.

Previously, the "docker port" command could be relied upon to produce
precisely one line of output. However, more recent versions appear to
produce two lines: one for IPv4, the other for IPv6.

This change makes the Address() method ignore the second IPv6 line,
and work only with the first IPv4 line.