Refactor: rename RunScriptWithEnv to RunScriptWithEnvGetStdout
...in ssh/remote_runner.go

In future commits, we will re-use the original name for a more
memory-efficient version of the function, which doesn't buffer all of
the StdOut.

The new RunScriptWithEnv will not return stdout the way
RunScriptWithEnvGetStdout does. All but one uses of
RunScriptWithEnvGetStdout throw that data away anyway, so they'll be
happy with the more efficient RunScriptWithEnv implementation.

This will mean means that if a backup script prints many gigabytes of
information to the stdout, our BBR CLI won't crash with an out of
memory error.