Refactor: rename internal methods
...in ssh/remote_runner.go

The function runOnInstanceWithLabel buffers stdout, so it can be
returned to the user. Stdout can get very large, and cause our program
to crash. Here we rename it to runOnInstanceWithLabelGetStdout, for
those occasions when we really need to get stdout.

In future commits, we will re-use the original name for a more
memory-efficient version of the function, which doesn't buffer all of
the stdout.