Use more memory-efficient RunScriptWithEnv method where possible
The new implementation of RunScriptWithEnv can be more
memory-efficient than the older RunScriptWithEnvGetStdout, because it
does not have to return the complete contents of stdout when it
returns.

Many uses of RunScriptWithEnvGetStdout immediately throw away the
value of stdout as soon as the method returns. In these cases it is
better to use RunScriptWithEnv, and not store the contents of stdout
in the first place.

Not all uses of RunScriptWithEnvGetStdout can be changed. In
particular, the use in instance/job_finder/findMetadata(...) requires
the data from stdout to discover metadata about the job to be
backed-up/restored.

The uses of RunScriptWithEnvGetStdout that call the actual backup and
restore scripts can be replaced with RunScriptWithEnv. In the event
that a backup or restore script prints enormous quantities of data to
stdout, this means the BBR CLI will no-longer run out of memory and
panic.