Make ssh tests run serially
The SSH tests make use of a local docker container to check that we
can perform certain operations over an SSH connection. I think that
container can become a source of test polution if the ssh tests are
run in parallel.

When run in parallel, on my machine, I observe approximately 1 failure
in every 5 or so test runs.

After making this makefile change so that they run serially, I was
just able to run the tests 10 times in a row with no failures.