Begin refactor: change signature of remote_runner.RunScriptWithEnv
This commit does not change any behaviour.

We have observed that, if a backup or restore script echos a great
deal of information to stdout, the method
remote_runner.RunScriptWithEnv can run out of memory and crash the BBR
CLI. This is because this method caches the entire output of the
command in memory, and returns it in a single string.

We want to change this method to accept an io.Writer instead. The
writer can be used to pass the contents of stdout directly to wherever
it needs to go as a stream, rather than caching in a single in-memory
string variable.

This commit begins that signature change. We introduce the writer, but
have not yet removed the return value. The writer is not used yet, we
have only made the types line up, and the tests pass.

This commit therefore contains the following changes:

- The signature of the concrete remote_runner.RunScriptWithEnv method
  now contains a new parameter for the writer.

- The signature of the method in the RemoteRunner interface matches.

- All callers of RunScriptWithEnv pass io.Discard for the new
  parameter.

- All test-fakes in the ssh package have been regenerated -- including
  FakeRemoteRunner, which forced the change.

- All tests which use the FakeRemoteRunner now explicitly ignore any
  values of the new Writer parameter to RunScriptWithEnv