Refactor: use r.connection.Stream instead of .Run
The r.connection.Run method returns the entirel contents of stdout in
a single string. The Stream method takes a Writer, and streams chunks
of stdout to it.

This refactor is about fixing an out-of-memory error caused by this
in-memory caching of the contents of stdout. This step doesn't fix the
whole problem, because we still cache the contents in this
method (RunScriptWithEnv). However, the problem is now more localised.