Refactor: inline logAndCheckErrors in RunScriptWithEnv
We want to stop buffering all of stdout in RunScriptWithEnv, so as to
avoid an out of memory crash.

We couldn't do that until now, because we called logAndCheckErrors,
which requires that we pass it a string containing everything from
stdout.

This commit inlines logAndCheckErrors, so that we can refactor away
the need to buffer all of stdout while logging.