Refactor: handle stdout logging incrementally in RunScriptWithEnv
This refactor is about fixing an out-of-memory crash, which results
from caching stdout in RunScriptWithEnv.

This commit fixes part of the problem. Previously we waited until the
remote command had finished, and then logged the entire contents of
stdout. Now we stream stdout into the log as it arrives.

We still cache all of stderr and log that at the end. We haven't
observed out-of-memory crashes from stdout in the wild, and we don't
yet have an upstream method for streaming it instead of caching
it. For now, just streaming stdout seems like a reasonable compromise.

Note that by streaming stdout and caching stderr, lines logged to each
of those file descriptors will appear out-of-order in our
logs. However, this was already the case, since after caching they
were logged serially.