Refactor: simplify RunScriptWithEnv logging
We previously inlined this logging code from the logAndCheckErrors
method, which is sometimes called with an empty 'label' argument.

However, RunScriptWithEnv is never called with an empty 'label'
argument, so we don't need to handle that special case to maintain
any backwards compatibility.

In the event that some future change results in RunScriptWithEnv being
called with an empty 'label' argument, the result will only be that
the log shows the label as explicitly empty, rather than missing
entirely. This actually seems preferable to me than having a
special-case, so long as there are no backwards-compatibility issues.