Refactor: make RunScriptWithEnv stream stdout to a Writer
This refactor is about fixing an out-of-memory crash, which is due to
RunScriptWithEnv caching stdout in memory. We want to stream it using
io.Writers instead.

Previous commits changed the signature of RunScriptWithEnv so that it
takes a Writer which we can stream to. This commit makes use of that
Writer, and checks that it works in the unit tests.

We have not yet removed the cached stdout value, which is still
returned by this function. That will be safe to remove once we've
changed all the callers of RunScriptWithEnv to use the Writer instead
of the return value.