Refactor: switch job_finder.findMetadata to use the Writer
This refactor is about fixing an out-of-memory crash that was caused
by stdout being cached in remote_runner.RunScriptWithEnv.

In previous commits, we set up RunScriptWithEnv so that it could
stream stdout to a Writer that it now takes as a parameter. In this
commit we change job_finder.findMetadata so that:

- it no longer uses the return value of RunScriptWithEnv
- instead it uses the Writer parameter to stream stdout into a buffer

Note that even if we remove all stdout caching from RunScriptWithEnv,
this findMetadata method will still cache stdout in the buffer
introduced in this commit. I believe this is fine, because
findMetadata expects specific bounded output from stdout. This is in
contrast to the occasions when we call backup and restore scripts, in
which the script output is unbounded.