Refactor: do not cache stdout in RunScriptWithEnv
This commit is the core of this refactor. While RunScriptWithEnv was
caching all of stdout in memory, we sometimes saw the CLI run out of
memory and crash.

In previous commits we introduced a Writer with which we could stream
stdout incrementally instead of caching it. In this commit we finally
remove the now-redundant cache.