Finish refactor: change signature of remote_runner.RunScriptWithEnv
This commit should not change any behaviour -- only type signatures.

In this refactor we have changed RunScriptWithEnv so that it no-longer
caches the contents of stdout in memory. This means extremely large
output to stdout should no longer cause the CLI to crash with an
out-of-memory error.

In this commit we make the final change to the signature of
RunScriptWithEnv. Since we no-longer cache stdout, we cannot return it
as a return value. Thanks to previous commits, that return value was
always empty anyway.

This commit therefore makes the following changes:

- The ssh.RemoteRunner interface changes so that RunScriptWithEnv and
  RunScript don't return a string

- The implementations of RunScriptWithEnv and RunScript change to
  match the interface.

- We re-generate the FakeRemoteRunner to match the interface.

- Calls to RunScriptWithEnv in remote_runner_test no longer have to
  explicitly discard its return value.

- Calls to RunScriptWithEnv and RunScript in instance/job.go and
  instance/job_finder.go no longer need to explicitly discard their
  return values.

- Assignments to FakeRemoteRunner.RunScriptWithEnvStub now need a
  different type signature. No behaviour has been changed, other than
  deleting unused dummy values.

- Calls to FakeRemoteRunner.RunScriptWithEnvReturns no longer need to
  take an unused string. These strings have been removed.