Delete the certificate copied from the golang testcert package
It looks like the cert previously included in fixtures/test.crt had
been copied from the internals of the golang net/http package[1]

So long as the golang maintainers never re-genreated that certificate,
all the tests that relied on our copy of the certificate continued to
pass. However, between go 1.17.6 and 1.17.7, the upstream certificate
was regenerated, and many of our tests failed.

In this commit, we delete the copied certificate, and instead use the
interface of the net/httptest package to get a copy of whatever
certificate is used in whatever go runtime we happen to be using at
the time.

This means our tests now pass with all supported versions of go,
regardless of which certificate they use internally.

[1] https://github.com/golang/go/blob/d17b65ff54a1824288eb68fe3fbc8c7beed14bb6/src/net/http/internal/testcert/testcert.go