Make tests more robust (#431)
* Make integration tests work with recent docker versions

The function mockInstance.Address() returns the address that you can
use to SSH into a particular docker container. To do this, it needs to
discover the port on your local machine that docker has mapped to port
22 inside the container.

The "docker port" command provides this information.

Previously, the "docker port" command could be relied upon to produce
precisely one line of output. However, more recent versions appear to
produce two lines: one for IPv4, the other for IPv6.

This change makes the Address() method ignore the second IPv6 line,
and work only with the first IPv4 line.

* Make integration tests more robust to varying SSH errors

Previously, the tests assumed that if SSH is unable to connect to a
given address, it would say "no such host". It seems that some SSH
implementations say "No address associated with hostname" instead.

We should accept either.

* Make ssh tests run serially

The SSH tests make use of a local docker container to check that we
can perform certain operations over an SSH connection. I think that
container can become a source of test polution if the ssh tests are
run in parallel.

When run in parallel, on my machine, I observe approximately 1 failure
in every 5 or so test runs.

After making this makefile change so that they run serially, I was
just able to run the tests 10 times in a row with no failures.

* Delete the certificate copied from the golang testcert package

It looks like the cert previously included in fixtures/test.crt had
been copied from the internals of the golang net/http package[1]

So long as the golang maintainers never re-genreated that certificate,
all the tests that relied on our copy of the certificate continued to
pass. However, between go 1.17.6 and 1.17.7, the upstream certificate
was regenerated, and many of our tests failed.

In this commit, we delete the copied certificate, and instead use the
interface of the net/httptest package to get a copy of whatever
certificate is used in whatever go runtime we happen to be using at
the time.

This means our tests now pass with all supported versions of go,
regardless of which certificate they use internally.

[1] https://github.com/golang/go/blob/d17b65ff54a1824288eb68fe3fbc8c7beed14bb6/src/net/http/internal/testcert/testcert.go