Make BBR more robust to large stdout streams. (#436)
* Begin refactor: change signature of remote_runner.RunScriptWithEnv

This commit does not change any behaviour.

We have observed that, if a backup or restore script echos a great
deal of information to stdout, the method
remote_runner.RunScriptWithEnv can run out of memory and crash the BBR
CLI. This is because this method caches the entire output of the
command in memory, and returns it in a single string.

We want to change this method to accept an io.Writer instead. The
writer can be used to pass the contents of stdout directly to wherever
it needs to go as a stream, rather than caching in a single in-memory
string variable.

This commit begins that signature change. We introduce the writer, but
have not yet removed the return value. The writer is not used yet, we
have only made the types line up, and the tests pass.

This commit therefore contains the following changes:

- The signature of the concrete remote_runner.RunScriptWithEnv method
  now contains a new parameter for the writer.

- The signature of the method in the RemoteRunner interface matches.

- All callers of RunScriptWithEnv pass io.Discard for the new
  parameter.

- All test-fakes in the ssh package have been regenerated -- including
  FakeRemoteRunner, which forced the change.

- All tests which use the FakeRemoteRunner now explicitly ignore any
  values of the new Writer parameter to RunScriptWithEnv

* Refactor: inline remote_runner.runOnInstanceWithLabel

If we want to be able to use the io.Writer argument "stdout", then
we'll need to be able to mess with the code that actually calls the
ssh connection.

Previously that code was hidden in the runOnInstanceWithLabel private
method. Now it's inline, so we can mess with it.

* Refactor: use r.connection.Stream instead of .Run

The r.connection.Run method returns the entirel contents of stdout in
a single string. The Stream method takes a Writer, and streams chunks
of stdout to it.

This refactor is about fixing an out-of-memory error caused by this
in-memory caching of the contents of stdout. This step doesn't fix the
whole problem, because we still cache the contents in this
method (RunScriptWithEnv). However, the problem is now more localised.

* Refactor: inline logAndCheckErrors in RunScriptWithEnv

We want to stop buffering all of stdout in RunScriptWithEnv, so as to
avoid an out of memory crash.

We couldn't do that until now, because we called logAndCheckErrors,
which requires that we pass it a string containing everything from
stdout.

This commit inlines logAndCheckErrors, so that we can refactor away
the need to buffer all of stdout while logging.

* Refactor: handle stdout logging incrementally in RunScriptWithEnv

This refactor is about fixing an out-of-memory crash, which results
from caching stdout in RunScriptWithEnv.

This commit fixes part of the problem. Previously we waited until the
remote command had finished, and then logged the entire contents of
stdout. Now we stream stdout into the log as it arrives.

We still cache all of stderr and log that at the end. We haven't
observed out-of-memory crashes from stdout in the wild, and we don't
yet have an upstream method for streaming it instead of caching
it. For now, just streaming stdout seems like a reasonable compromise.

Note that by streaming stdout and caching stderr, lines logged to each
of those file descriptors will appear out-of-order in our
logs. However, this was already the case, since after caching they
were logged serially.

* Refactor: remove runOnInstanceWithLabel

Now that we no-longer use this method from RunScriptWithEnv, it
doesn't need to be separate from runOnInstance any more.

* Refactor: simplify RunScriptWithEnv logging

We previously inlined this logging code from the logAndCheckErrors
method, which is sometimes called with an empty 'label' argument.

However, RunScriptWithEnv is never called with an empty 'label'
argument, so we don't need to handle that special case to maintain
any backwards compatibility.

In the event that some future change results in RunScriptWithEnv being
called with an empty 'label' argument, the result will only be that
the log shows the label as explicitly empty, rather than missing
entirely. This actually seems preferable to me than having a
special-case, so long as there are no backwards-compatibility issues.

* Refactor: make RunScriptWithEnv stream stdout to a Writer

This refactor is about fixing an out-of-memory crash, which is due to
RunScriptWithEnv caching stdout in memory. We want to stream it using
io.Writers instead.

Previous commits changed the signature of RunScriptWithEnv so that it
takes a Writer which we can stream to. This commit makes use of that
Writer, and checks that it works in the unit tests.

We have not yet removed the cached stdout value, which is still
returned by this function. That will be safe to remove once we've
changed all the callers of RunScriptWithEnv to use the Writer instead
of the return value.

* Refactor: switch job_finder.findMetadata to use the Writer

This refactor is about fixing an out-of-memory crash that was caused
by stdout being cached in remote_runner.RunScriptWithEnv.

In previous commits, we set up RunScriptWithEnv so that it could
stream stdout to a Writer that it now takes as a parameter. In this
commit we change job_finder.findMetadata so that:

- it no longer uses the return value of RunScriptWithEnv
- instead it uses the Writer parameter to stream stdout into a buffer

Note that even if we remove all stdout caching from RunScriptWithEnv,
this findMetadata method will still cache stdout in the buffer
introduced in this commit. I believe this is fine, because
findMetadata expects specific bounded output from stdout. This is in
contrast to the occasions when we call backup and restore scripts, in
which the script output is unbounded.

* Refactor: do not cache stdout in RunScriptWithEnv

This commit is the core of this refactor. While RunScriptWithEnv was
caching all of stdout in memory, we sometimes saw the CLI run out of
memory and crash.

In previous commits we introduced a Writer with which we could stream
stdout incrementally instead of caching it. In this commit we finally
remove the now-redundant cache.

* Finish refactor: change signature of remote_runner.RunScriptWithEnv

This commit should not change any behaviour -- only type signatures.

In this refactor we have changed RunScriptWithEnv so that it no-longer
caches the contents of stdout in memory. This means extremely large
output to stdout should no longer cause the CLI to crash with an
out-of-memory error.

In this commit we make the final change to the signature of
RunScriptWithEnv. Since we no-longer cache stdout, we cannot return it
as a return value. Thanks to previous commits, that return value was
always empty anyway.

This commit therefore makes the following changes:

- The ssh.RemoteRunner interface changes so that RunScriptWithEnv and
  RunScript don't return a string

- The implementations of RunScriptWithEnv and RunScript change to
  match the interface.

- We re-generate the FakeRemoteRunner to match the interface.

- Calls to RunScriptWithEnv in remote_runner_test no longer have to
  explicitly discard its return value.

- Calls to RunScriptWithEnv and RunScript in instance/job.go and
  instance/job_finder.go no longer need to explicitly discard their
  return values.

- Assignments to FakeRemoteRunner.RunScriptWithEnvStub now need a
  different type signature. No behaviour has been changed, other than
  deleting unused dummy values.

- Calls to FakeRemoteRunner.RunScriptWithEnvReturns no longer need to
  take an unused string. These strings have been removed.