Harden tests by not relying on implemtation details
[#182248334]

Invoking ParseAuthorizedKey in tests betrays the purpose of
testing by harnessing some implementation knowledge to make the
test pass.

Since the key used in tests uses RSA format, we are now making
that explicit in our variables names. Moreover, since the key
format is know so is the expected SSH algorithm: `ssh-rsa`.