Update expected behaviour. Tests not passing after this change.
[#182248334]

When an RSA key is used for the SSH connection we want BBR to
list the three algorithms supported by the crypto/ssh package
and not only `ssh-rsa`.

`ssh-rsa` algorithm has been deprecated in newer versions of
OpenSSH: https://www.openssh.com/txt/release-8.7

As specified in the deprecation notice:
> deactivation of "ssh-rsa" signatures does not necessarily
require cessation of use for RSA keys. In the SSH protocol, keys may be
capable of signing using multiple algorithms. In particular, "ssh-rsa"
keys are capable of signing using "rsa-sha2-256" (RSA/SHA256),
"rsa-sha2-512" (RSA/SHA512) and "ssh-rsa" (RSA/SHA1). Only the last of
these is being turned off by default.

Is precisely because of this that we are changing our tests
to require that we list the three algorithsm when an RSA key
is used.