Fix logic to determine supportedEncryptionAlgorithm
[#182248334]

This logic treats RSA keys as a special case and lists three
algorithms known to be supported by the time of this commit.

While for any other NON RSA key, the format of the key is used
as the name for the algorithm itself. This is an acceptable and
common practice in the crypto key formats/algorithms ecosystem.
With the aforementioned RSA being an exception.