182248334 make ssh handshake use updated algorithms (#541)
* Revert "Fix ssh issues with Windows Stemcells 2019.46+ (#539)"

This reverts commit e5bd7d42cfa7b83c7b5d5ed0f4ba4d776cdf7e1c.

* Extract supportedEncryptionAlgorithms so we can add logic to it

[#182248334]

* Harden tests by not relying on implemtation details

[#182248334]

Invoking ParseAuthorizedKey in tests betrays the purpose of
testing by harnessing some implementation knowledge to make the
test pass.

Since the key used in tests uses RSA format, we are now making
that explicit in our variables names. Moreover, since the key
format is know so is the expected SSH algorithm: `ssh-rsa`.

* Update expected behaviour. Tests not passing after this change.

[#182248334]

When an RSA key is used for the SSH connection we want BBR to
list the three algorithms supported by the crypto/ssh package
and not only `ssh-rsa`.

`ssh-rsa` algorithm has been deprecated in newer versions of
OpenSSH: https://www.openssh.com/txt/release-8.7

As specified in the deprecation notice:
> deactivation of "ssh-rsa" signatures does not necessarily
require cessation of use for RSA keys. In the SSH protocol, keys may be
capable of signing using multiple algorithms. In particular, "ssh-rsa"
keys are capable of signing using "rsa-sha2-256" (RSA/SHA256),
"rsa-sha2-512" (RSA/SHA512) and "ssh-rsa" (RSA/SHA1). Only the last of
these is being turned off by default.

Is precisely because of this that we are changing our tests
to require that we list the three algorithsm when an RSA key
is used.

* Support `rsa-sha2-256` and `rsa-sha2-512` algorithms

[#182248334]

* Duplicate client_test but now testing with a ECDSA public key

[#182248334]

Tests are passing but totally misleading. We have created an
IDENTICAL copy of the client_test file. In next commits we are
going to introduce changes to really test using an ECDSA key.

* Account for situations in which non RSA keys are used

[#182248334]

Tests are not passing after this commit.
Tests have been updated to make explicit that for non RSA keys
we want to list a single algorithm named after the detected
format used to encode the key.

* Fix logic to determine supportedEncryptionAlgorithm

[#182248334]

This logic treats RSA keys as a special case and lists three
algorithms known to be supported by the time of this commit.

While for any other NON RSA key, the format of the key is used
as the name for the algorithm itself. This is an acceptable and
common practice in the crypto key formats/algorithms ecosystem.
With the aforementioned RSA being an exception.

* Run `go fmt`

* De-duplicate the ECDSA tests

Rather than duplicating all the bosh client tests, we can make do with
just one. When a server is configured with an ECDSA public key, we
check that the handshake method used is the ecdsa-sha2-nistp521 one.

[#182248334]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

* Improve readability of supportedEncryptionAlgorithms function

This commit doesn't change the behaviour of the code, because the
changed line only ever executes if key.Type is equal to
gossh.KeyAlgoRSA already.

This way, we're clearer about our intent. This is the RSA branch.

[#182248334]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

Co-authored-by: Fernando Naranjo <fnaranjo@vmware.com>