Ensure fields have the right format
[#182318242]

Ensure ca_cert and jumpbox_ssh_key are encoded as multiline block strings in yaml.
To do this we must ensure the certificate and private key are formatted using newlines
instead of as a single-line string separating sections by spaces.

Ensure jumpbox_url includes a port number. Default to 22 if not port was specified.