Rewrite bbr-unit test to work with a local docker_host (#747)
* Rewrite bbr-unit test to work with a local docker_host

[#183300237]

Previously, the test relied on an external docker_host deployed to GCP.
That environment got deleted or corrupted at some point which was causing
the pipeline to fail.

This change removes the dependency on an external docker_host by running
docker-in-docker using docker-compose. This way, the test can be run locally
or in GitHub Actions seemlessly.

* Fix some tests which only fail when run in GHActions

[#183300237]