Fix evergreen s3-validator pipeline
We're seeing some CI runs like this[1] which ought to fail, but don't.
This is because CI is currently only tracking whether or not our secrets
redactor succeeds, not whether the underlying command succeeds.

We can fix this by using `set -o pipefail` at the top of the bash
script.

[1] https://ci.cryo.cf-app.com/teams/bosh-backup-restore/pipelines/bbr-cli/jobs/test-build-s3-config-validator/builds/841