Fix evergreen s3-validator pipeline (#764)
* Fix evergreen s3-validator pipeline

We're seeing some CI runs like this[1] which ought to fail, but don't.
This is because CI is currently only tracking whether or not our secrets
redactor succeeds, not whether the underlying command succeeds.

We can fix this by using `set -o pipefail` at the top of the bash
script.

[1] https://ci.cryo.cf-app.com/teams/bosh-backup-restore/pipelines/bbr-cli/jobs/test-build-s3-config-validator/builds/841

* Increate run-b-drats timeout

- This is not directly related to this PR but currently the pipeline is
  failing on BBR timing out
- We suspect this is related the recent changes in our IaaS resource
  config in order to cut costs

* Rename job

- Also  not related to this PR but the current job does not do what it
  says which makes things more complicated

Co-authored-by: fnaranjo <fnaranjo@vmware.com>
Co-authored-by: Diego Lemos <dlemos@vmware.com>