[s3 validator] ignore creds if we're using IAM
We can authenticate to S3 in one of two ways:
- With an ID and a secret key
- Using IAM instance profiles

The former case is more usual. We include creds in buckets.json.

The latter case works if the backupper is being run on an EC2 instance,
and if that instance is already pre-authorised with S3. In that case, no
additional credentials are needed.

Previously we assumed that all users would always use an ID and secret
access key. Now we will not bail out (claiming an invalid config) if the
user is properly set up to use instance profiles.

[#184963912]