fix: Make validator work with instance profiles (#915)
* [s3 validator] ignore creds if we're using IAM

We can authenticate to S3 in one of two ways:
- With an ID and a secret key
- Using IAM instance profiles

The former case is more usual. We include creds in buckets.json.

The latter case works if the backupper is being run on an EC2 instance,
and if that instance is already pre-authorised with S3. In that case, no
additional credentials are needed.

Previously we assumed that all users would always use an ID and secret
access key. Now we will not bail out (claiming an invalid config) if the
user is properly set up to use instance profiles.

[#184963912]

* [s3 validator] check for duplicate creds

If we intend to use instance profiles to authenticate with S3, then we
shouldn't be passing around secret access keys. They won't be used, and
we shouldn't risk leaking them if some vulnerability or other crops up.

[#184963912]

* [s3 validator] Backfill unversioned IAMprofile test

In the previous commit we introduced a test like this to drive out the
check that IAM profile creds were handled correctly. This new test
checks that the new check works with unversioned bucket configs as well
as with versioned ones.

[#184963912]

* [s3 validator] check for missing backup buckets

When asking for an unversioned backup, we should check that the
buckets.json specified backup buckets.

[#184963912]

* [s3 validator] clarify IAM config error message

[#184963912]

* [s3 validator] describe all the issues we can

Don't bail out as soon as we find one issue. Keep going as long as
possible, and describe all the issues we can find.

[#184963912]

* [ci validator] make printer tests IAM-aware

[#184963912]

* [s3 validator] test out IAM profile auth

This isn't wired in yet, but the functionality is now available in our
internal s3 client.

[#184963912]

* [s3 validator] wire in the new IAM profile auth

As of this commit, we respect the buckets.json when it asks us to use
IAM Profiles instead of secrets for auth.

The functionality was tested out in previous commits. This one just
joins up the wires between the config and the client-construction code.

[#184963912]

* [s3 validator] Clean up

* [s3 validator] Use IAM profiles

* [s3 validator] Prevent endpoint being set when IAM profile is used

[#184963912]

Signed-off-by: Konstantin Semenov <ksemenov@vmware.com>

---------

Signed-off-by: Konstantin Semenov <ksemenov@vmware.com>
Co-authored-by: Gareth Smith <sgareth@vmware.com>