updating the image broke tests
the error msg that the docker container returns on an unresolvable
host changed when we switched to the go.1.21 image. These tests expect
an DNS issue and have been updated before to accomodate the behavior
when running in docker. Running locally these tests still pass and it
seems that the only thing that really changed is the msg that the failed
dns lookup produces.

see 815a457d382ab3bae196aaa100a7d62bb5ca06c6 for a similar change to these
tests when they were moved to be run in a docker in docker setup.