reactivate s3 config validator tests
[#185999452]

this is required because of the move to cloudgate. the system blobst supports
two methods for authenticating s3:
- iam profiles
- access keys

The current setup uses gcp envs from toolsmiths to avoid managing snowflake envs.
we cannot use iam profiles from gcp.

Cloudgate doesn't allow iam users in the actual account. we need to assume role.
this is not, and most probably wont be, supported in CC blobstore because there's
no usecase for it.

We still want to test the s3 config validator, to re enable these tests. The support
for switch role was added a while ago and this switches over to using it.

- add go mod vendor:
  fixes:

  ```
  vendor/github.com/onsi/ginkgo/v2/ginkgo/main.go:7:2: cannot find module providing package github.com/onsi/ginkgo/v2/ginkgo/build: import lookup disabled by -mod=vendor
          (Go version in go.mod is at least 1.14 and vendor directory exists.)
  ```

- switch image to cryo/essentials
- remove claiming env
  run tests in container directly
- update task to render the required config file in the location where the validator
  binary will look for it

- Remove setting Endpoint in case Role is used. The line threw a warning:
  ```
  options.EndpointResolver is deprecated: Deprecated: EndpointResolver and WithEndpointResolver. Providing a value for this field will likely prevent you from using any endpoint-related service features released after the introduction of EndpointResolverV2 and BaseEndpoint.
  ```
  Since we do not rely on custom endpoints in testing envs and the switch role support is purely
  for fullfilling requirements on our testing envs as of now, this is removed as the
  code failed with:
  ```
   failed to sign request: failed to retrieve credentials: operation error STS: AssumeRole, https response error StatusCode: 400, RequestID: CEGS9A525HE344HW, api error UnknownError: UnknownError]
  ```
  on all operations