avoid creating firewall rule
[ #186313155 ]

to run the bbr test suite we deploy a custom jumpbox. that jumpbox is
using the `jumpbox` network tag to whitelist specific ports between
itself and other vms.

the firewall rule behind this tag has been created by the removed task.

we want to move this CI to utilize shepherd environments to avoid managing
infrastructure and that implies that we do not have access to the underlying
gcp account and thus cannot create any firewall rules via gcloud cmd.

The rule that previously was created is the same rule that the actual jumpbox
of the environment uses albeit exposed via a different tag. Since the underlying
cfd is created via bbl, we can compute the relevant firewall rule name. it will
be `$ENV_NAME-bosh-open`. The easiest way to apply this network tag to the jumpbox
vm that the test runs create, is to update the vm_type that the jumpbox
instance_group would use. This is hardcoded into the fixture located at
`./fixtures/jumpbox.yml`.