add a go mod tidy before go mod vendor
It’s generally a good practice to add a call to go mod tidy before running go mod vendor.
This ensures that your go.mod file is up-to-date and accurately reflects the dependencies
before running the tests.
Not calling go mod tidy caused an issue in one of the dependencies updates because go version
needed to be changed to 1.23 from 1.22